Convert getFamily to TS (#463)
